Fix setup entry signature by forwarding to sensor platform
